LogToSplunk attempts to reconnect to splunk after splunk restart will continue sending data to available splunks and will store data to send to splunks that are unavailable.
Time between attempts to reconnect to a splunk instance can be set in
config.

Specification of wood types by LogToSplunk can be turned on and off in
config-code for getting specific wood types like 'spruce' or 'birch' may
be unreliable in the future.

Issues:
-may want to adjust how connection information is reported -
specifically the suggestion that the splunk host:port has not been set
in config should only show up on initialization.

-need to test running with multiple splunk instances